+++ gateway/config/development.lua
@@ -2,4 +2,6 @@ return {
     worker_processes = '1',
     master_process = 'off',
     lua_code_cache = 'off',
+    configuration_loader = 'lazy',
+    configuration_cache = 0,
 }