+++ gateway/src/apicast/configuration_loader.lua
@@ -122,7 +122,7 @@ function boot.init(configuration)
   if config and init then
     ngx.log(ngx.DEBUG, 'downloaded configuration: ', config)
   else
-    ngx.log(ngx.EMERG, 'failed to load configuration, exiting (code ', code, ')\n',  err)
+    ngx.log(ngx.EMERG, 'failed to load configuration, exiting (code ', code, ')\n',  err or ngx.config.debug and debug.traceback())
     os.exit(1)
   end
 
@@ -213,8 +213,10 @@ function lazy.rewrite(configuration, host)
   return configuration
 end
 
+local test = { init = noop, init_worker = noop, rewrite = noop }
+
 local modes = {
-  boot = boot, lazy = lazy, default = 'lazy'
+  boot = boot, lazy = lazy, default = 'lazy', test = test
 }
 
 function _M.new(mode)